player abstracted to have an ai player

+++ lib/screens/game.dart
@@ -3,13 +3,10 @@ import 'package:assets_audio_player/assets_audio_player.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/widgets.dart';
 import 'package:stanley/stanley.dart';
-import 'package:tetris/ai/ai.dart';
-import 'package:tetris/ai/dumb.dart';
-import 'package:tetris/ai/force.dart';
-
 import 'package:tetris/model/game.dart';
 import 'package:tetris/model/stats.dart';
 import 'package:tetris/model/tetromino.dart';
+import 'package:tetris/utils/player.dart';
 import 'package:tetris/utils/ui.dart';
 import 'package:tetris/widgets/block_painter.dart';
 import 'package:tetris/widgets/board_widget.dart';
@@ -19,31 +16,32 @@ class GameScreen extends StatefulWidget {
   _GameScreenState createState() => _GameScreenState();
 }
 
-class _GameScreenState extends State<GameScreen> {
+class _GameScreenState extends State<GameScreen> implements TetrisUI {
   static const int kTickAiDuration = 100;
 
   Game _game;
   bool _vertDragConsumed = false;
   double _horizDragDelta = 0;
   AssetsAudioPlayer _assetsAudioPlayer;
-  Pajitnov _ai;
+  Player _player;
 
   @override
   void initState() {
     super.initState();
-    _ai = new BruteForce();
+    _player = new AiPlayer(ui: this);
     _assetsAudioPlayer = AssetsAudioPlayer();
     _assetsAudioPlayer.loop = true;
     _assetsAudioPlayer.open(
       Audio('assets/sounds/theme.mp3'),
       volume: 0.5,
     );
-    this.reset();
+    this.restartGame();
   }
 
-  reset() {
+  @override
+  restartGame() {
     _game = Game();
-    this._tick(true);
+    _player.startGame(_game);
   }
 
   @override
@@ -96,30 +94,34 @@ class _GameScreenState extends State<GameScreen> {
             ),
             GestureDetector(
               onHorizontalDragUpdate: (DragUpdateDetails details) {
-                _horizDragDelta += details.delta.dx;
-                if (_horizDragDelta < -10) {
-                  _game.moveLeft();
-                  setState(() {});
-                  _horizDragDelta = 0;
-                } else if (_horizDragDelta > 10) {
-                  _game.moveRight();
-                  setState(() {});
-                  _horizDragDelta = 0;
+                if (_player.userCanInteract) {
+                  _horizDragDelta += details.delta.dx;
+                  if (_horizDragDelta < -10) {
+                    _game.moveLeft();
+                    setState(() {});
+                    _horizDragDelta = 0;
+                  } else if (_horizDragDelta > 10) {
+                    _game.moveRight();
+                    setState(() {});
+                    _horizDragDelta = 0;
+                  }
                 }
               },
               onHorizontalDragEnd: (_) => _horizDragDelta = 0,
               onHorizontalDragCancel: () => _horizDragDelta = 0,
               onVerticalDragStart: (_) => _vertDragConsumed = false,
               onVerticalDragUpdate: (DragUpdateDetails details) {
-                if (_vertDragConsumed == false) {
-                  if (details.delta.dy < 0) {
-                    _game.rotate();
-                    setState(() {});
-                    _vertDragConsumed = true;
-                  } else if (details.delta.dy > 0) {
-                    _game.drop();
-                    setState(() {});
-                    _vertDragConsumed = true;
+                if (_player.userCanInteract) {
+                  if (_vertDragConsumed == false) {
+                    if (details.delta.dy < 0) {
+                      _game.rotate();
+                      setState(() {});
+                      _vertDragConsumed = true;
+                    } else if (details.delta.dy > 0) {
+                      _game.drop();
+                      setState(() {});
+                      _vertDragConsumed = true;
+                    }
                   }
                 }
               },
@@ -130,8 +132,10 @@ class _GameScreenState extends State<GameScreen> {
               ),
               onTap: () {
                 if (_game.isFinished) {
-                  this.reset();
-                } else {
+                  if (_player.autoRestart == false) {
+                    this.restartGame();
+                  }
+                } else if (_player.userCanInteract) {
                   _game.rotate();
                   setState(() {});
                 }
@@ -140,54 +144,40 @@ class _GameScreenState extends State<GameScreen> {
             SizedBox(
               height: 16,
             ),
-            Column(
+            Row(
+              mainAxisAlignment: MainAxisAlignment.spaceEvenly,
               crossAxisAlignment: CrossAxisAlignment.start,
               children: [
-                UIUtils.text('holes: ${stats.numHoles}'),
-                UIUtils.text('min: ${stats.minHeight}'),
-                UIUtils.text('max: ${stats.maxHeight}'),
-                UIUtils.text('avg: ${stats.avgHeight}'),
-                UIUtils.text('std: ${stats.heightSD}'),
+                Column(
+                  crossAxisAlignment: CrossAxisAlignment.start,
+                  children: [
+                    UIUtils.text('holes: ${stats.numHoles}'),
+                    UIUtils.text('min: ${stats.minHeight}'),
+                    UIUtils.text('max: ${stats.maxHeight}'),
+                    UIUtils.text('avg: ${stats.avgHeight}'),
+                    UIUtils.text('std: ${stats.heightSD}'),
+                  ],
+                ),
+                UIUtils.text(_player.getInfo() ?? ''),
               ],
-            )
+            ),
           ],
         ),
       ),
     );
   }
 
-  void _tick(bool setTimer) {
-    if (_game.tick()) {
-      _horizDragDelta = 0;
-      _vertDragConsumed = true;
-    }
-    setState(() {
-      if (_game.isFinished == false) {
-
-        // let ai play
-        if (_ai != null) {
-          _ai.play(_game, () {
-            //setState(() {});
-            //sleep(Duration(milliseconds: 50));
-          });
-        }
-
-        // set next timer
-        if (setTimer) {
-          Future.delayed(Duration(milliseconds: _ai == null ? _game.delay : kTickAiDuration), () {
-            _tick(true);
-          });
-        }
-
-      } else if (_ai != null) {
-
-        Future.delayed(Duration(milliseconds: 500), () {
-          //this.reset();
-        });
-
-      }
-    });
+  @override
+  void currentPieceDone() {
+    _horizDragDelta = 0;
+    _vertDragConsumed = true;
   }
+
+  @override
+  void stateUpdateNeeded() {
+    setState(() {});
+  }
+
 }
 
 class Score extends StatelessWidget {
